requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.6
django-admin-interface==0.26.1
django-advanced-password-validation==1.2.0
django-colorfield==0.10.1
django-model-utils==4.5.0
django-notifications-hq==1.8.3
djangorestframework==3.14.0
django-extensions==3.2.3
idna==3.7
jsonfield==3.1.0
locust==2.29.1
numpy==1.26.4
opencv-python==4.9.0.80
paydunya==1.0.7
pillow==10.3.0
pypng==0.20220715.0
python-slugify==8.0.4
pytz==2024.1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
sqlparse==0.5.0
swapper==1.3.0
text-unidecode==1.3
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
whitenoise==6.6.0
